disable tracking of the developer in AndroidJUnitRunner
The Android Testing Support Library collects analytics to continuously
improve the testing experience. More specifically, it uploads a hash of the
package name of the application under test for each invocation. If you do
not wish to upload this data, you can opt-out by passing the following
argument to the test runner: disableAnalytics "true".