default to broadcasting to the current app only, with opt-in global
Most apps that use TorService as a library do not want to share the tor
instance with other apps.  So TorService should not send global broadcasts
in that case.  Apps like Orbot and TorServices still need to send global
broadcasts, so there is a method to re-enable that.

As a follow-on to the issues raised in #71.