Fixes #67
Exceptions thrown on the ControlSocketThread could not unlock the lock set in the tor thread, so the service
stops and the lock was forever stuck since only the killed tor service was capable of unlocking it.

This commit makes it so that when either the "tor" or CST threads call stopSelf(), TorService becomes unlocked.

Also adds a few annotations, lint improvements, etc.